automatically retrieve shapefile from OGGM
