--- notebooks/example_data/iceland/glacier_outlines/06_rgi60_Iceland.prj
@@ -1,1 +1,0 @@
-GEOGCS["GCS_WGS_1984",DATUM["D_WGS_1984",SPHEROID["WGS_1984",6378137.0,298.257223563]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.0174532925199433]]
